Read startup arguments (url, command line arguments) from a file.
Crosswalk is expected to launch with command line parameters which will
tell the URL/Manifest file to render as well as all the options (like
debug, extensions and so forth). In an environment where Crosswalk is
expected to launch without command line parameters (such as in stores like
Windows Store) we read all these informations from a "startup.json" file.
This file is created by app-tools with the correct arguments specified
in manifest.json from the developers.